created new js file fix linter error
